Tests/LibWeb/Text/expected/canvas/drawimage-canvas-intermediate-state.txt

B sampled A's intermediate state (red): true
A has its final state (blue): true
